requirements.txt
fastapi
uvicorn
scikit-learn
tensorflow
tensorboard
pillow
requests
python-multipart